requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
attrs==23.2.0
blessed==1.20.0
blinker==1.8.2
cachetools==5.3.3
cement==2.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
docutils==0.16
easyocr==1.7.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
fonttools==4.53.0
fsspec==2024.6.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.64.1
gTTS==2.5.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
onnx==1.16.1
onnx-tf==1.5.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pathspec==0.10.1
pillow==10.3.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pygame==2.5.2
Pygments==2.18.0
pyparsing==3.1.2
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
rsa==4.7.2
s3transfer==0.10.2
scikit-image==0.24.0
scipy==1.13.1
seaborn==0.13.2
semantic-version==2.8.5
setuptools==70.1.1
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.36.0
sympy==1.12.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==1.1.0
tifffile==2024.6.18
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.40
ultralytics-thop==2.0.0
urllib3==1.26.19
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.1.9
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
